Add Enum reflection in generated protobuf code
Now the only generation change needed is enabling the custom transient
property.

It's looking like the only way to do that is have that protobuf
extension a dependency of the code generation. It would need to be
built into CPP code and into the generator directly. Otherwise, the
generator does not know that extension exists.